pamtalk-esg-chain/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
py-algorand-sdk==2.8.0
python-dotenv==1.0.1
pyteal==0.27.0
psycopg2-binary==2.9.7